Change namednativequeries with result set mappings to named queries with explicit constructors to fix jdbc errors with native image.
