+++ visa-core/src/main/java/eu/ill/visa/core/entity/partial/NumberInstancesByImage.java
@@ -1,16 +1,18 @@
-package eu.ill.visa.core.domain;
+package eu.ill.visa.core.entity.partial;
 
-public class NumberInstancesByFlavour {
+import io.quarkus.runtime.annotations.RegisterForReflection;
+
+@RegisterForReflection
+public class NumberInstancesByImage {
     private Long id;
     private String name;
+    private String version;
     private Long total ;
 
-    public NumberInstancesByFlavour() {
-    }
-
-    public NumberInstancesByFlavour(final Long id, final String name, final Long total) {
+    public NumberInstancesByImage(final Long id, final String name, final String version, final Long total) {
         this.id = id;
         this.name = name;
+        this.version = version;
         this.total = total;
     }
 
@@ -28,6 +30,14 @@ public class NumberInstancesByFlavour {
 
     public void setName(String name) {
         this.name = name;
+    }
+
+    public String getVersion() {
+        return version;
+    }
+
+    public void setVersion(String version) {
+        this.version = version;
     }
 
     public Long getTotal() {
