+++ visa-persistence/src/main/java/eu/ill/visa/persistence/repositories/InstanceRepository.java
@@ -6,6 +6,9 @@ import eu.ill.visa.core.entity.*;
 import eu.ill.visa.core.entity.enumerations.InstanceMemberRole;
 import eu.ill.visa.core.entity.enumerations.InstanceState;
 import eu.ill.visa.core.entity.partial.InstancePartial;
+import eu.ill.visa.core.entity.partial.NumberInstancesByCloudClient;
+import eu.ill.visa.core.entity.partial.NumberInstancesByFlavour;
+import eu.ill.visa.core.entity.partial.NumberInstancesByImage;
 import jakarta.inject.Inject;
 import jakarta.inject.Singleton;
 import jakarta.persistence.EntityManager;
