+++ visa-web-graphql/src/main/java/eu/ill/visa/web/graphql/types/NumberInstancesByFlavourType.java
@@ -1,6 +1,6 @@
 package eu.ill.visa.web.graphql.types;
 
-import eu.ill.visa.core.domain.NumberInstancesByFlavour;
+import eu.ill.visa.core.entity.partial.NumberInstancesByFlavour;
 import io.smallrye.graphql.api.AdaptToScalar;
 import io.smallrye.graphql.api.Scalar;
 import jakarta.validation.constraints.NotNull;
